Move FileLock to niflheim project
