+++ niflheim/src/main/scala/quasar/niflheim/CookStateLog.scala
@@ -16,8 +16,6 @@
 
 package quasar.niflheim
 
-import quasar.precog.util.FileLock
-
 import org.objectweb.howl.log._
 
 import org.slf4s.Logging
